requirements.txt
albumentations==1.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.0.0
cryptography==37.0.1
cycler==0.11.0
cytoolz==0.11.0
dask==2022.2.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.3.0
dill==0.3.5.1
docker-pycreds==0.4.0
easydict==1.9
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.1
filelock==3.8.0
fonttools==4.33.3
fsspec==2022.2.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.9.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
importlib-metadata==4.12.0
importlib-resources==5.2.0
ipykernel==6.9.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
kornia==0.6.4
locket==0.2.1
lpips==0.1.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multiprocess==0.70.13
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.7.1
notebook==6.4.11
numpy==1.21.5
opencv-python-headless==4.6.0.66
p-tqdm==1.4.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathos==0.2.9
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.1.2
pox==0.3.1
ppft==1.7.6.5
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
regex==2022.9.13
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.9.1
Send2Trash==1.8.0
sentry-sdk==1.8.0
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.7.2
tokenizers==0.12.1
toolz==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchfile==0.1.0
torchmetrics==0.9.2
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
transformers==4.21.3
typing_extensions==4.1.1
urllib3==1.26.9
visdom==0.1.8.9
wandb==0.12.21
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
wget==3.2
wheel==0.37.1
widgetsnbextension==3.6.1
zipp==3.8.0
